requirements.txt
asgiref==3.7.2
bcrypt==4.0.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
cryptography==41.0.3
defusedxml==0.7.1
diff-match-patch==20230430
Django==4.2.3
django-import-export==3.2.0
djangorestframework==3.14.0
et-xmlfile==1.1.0
idna==3.4
josepy==1.13.0
MarkupPy==1.14
mozilla-django-oidc==3.0.0
odfpy==1.4.1
openpyxl==3.1.2
Pillow==10.0.0
psycopg==3.1.10
pycparser==2.21
PyJWT==2.7.0
pyOpenSSL==23.2.0
pypng==0.20220715.0
pytz==2023.3
PyYAML==6.0.1
qrcode==7.4.2
requests==2.31.0
sqlparse==0.4.4
tablib==3.5.0
typing_extensions==4.7.1
urllib3==2.0.3
xlrd==2.0.1
xlwt==1.3.0
gunicorn==21.2.0